requirements.txt
# Core agent & toolchain
strands-agents==0.1.8
strands-agents-tools==0.1.6
mcp==1.9.4
uv==0.7.14

# AWS Bedrock LLM via Boto3 (for Claude)
boto3==1.38.42
botocore==1.38.42

# Optional but useful
rich==14.0.0  # Pretty terminal output
